moved wpcheck
