+++ src/lib/wpcheck/types.ts
@@ -1,0 +1,12 @@
+export enum Severity {
+  Good,
+  Med,
+  Bad,
+}
+
+export type Fault = {
+  offenderMission?: string,
+  offenderIndex?: number,
+  message: string,
+  severity: Severity
+}
